// src/app/models/user.ts
export interface IUser {
  id: number;
  firstname: string;
  lastname: string;
  username: string;
  email: string;
  email_verified_at: string;
  phone: string;
  gender: string;
  picture: string;
  email_token: number;
  is_active: number;
  deleted_at: string;
  remember_token: string;
  type: string;
  pin: string;
  referral_code: string;
  unique_code: string;
  created_at: string;
  updated_at: string;
  address: string;
  middlename: string;
  is_email_verify: number;
  is_kyc1_verify: number;
  is_kyc2_verify: number;
}